feat: component support moveTo
