+++ packages/vstory/src/player/processor/processorMap.ts
@@ -7,7 +7,7 @@ import { RectVisibilityActionProcessor } from './component/rect/rect-visibility'
 import { ShapeVisibilityActionProcessor } from './component/shape/shape-visibility';
 import { TextVisibilityActionProcessor } from './component/text/text-visibility';
 import { TimelineVisibilityActionProcessor } from './component/timeline/timeline-visibility';
-import { CommonStyleActionProcessor } from './component/common-component';
+import { CommonMoveToActionProcessor, CommonStyleActionProcessor } from './component/common-component';
 
 export const processorChartMap = {
   [StoryChartType.VCHART]: {
@@ -20,32 +20,38 @@ export const processorComponentMap = {
   [StoryComponentType.RECT]: {
     [ACTION_TYPE.APPEAR]: new RectVisibilityActionProcessor(),
     [ACTION_TYPE.DISAPPEAR]: new RectVisibilityActionProcessor(),
-    [ACTION_TYPE.STYLE]: new CommonStyleActionProcessor()
+    [ACTION_TYPE.STYLE]: new CommonStyleActionProcessor(),
+    [ACTION_TYPE.MOVETO]: new CommonMoveToActionProcessor()
   },
   [StoryComponentType.LINE]: {
     [ACTION_TYPE.APPEAR]: new LineVisibilityActionProcessor(),
     [ACTION_TYPE.DISAPPEAR]: new LineVisibilityActionProcessor(),
-    [ACTION_TYPE.STYLE]: new CommonStyleActionProcessor()
+    [ACTION_TYPE.STYLE]: new CommonStyleActionProcessor(),
+    [ACTION_TYPE.MOVETO]: new CommonMoveToActionProcessor()
   },
   [StoryComponentType.SHAPE]: {
     [ACTION_TYPE.APPEAR]: new ShapeVisibilityActionProcessor(),
     [ACTION_TYPE.DISAPPEAR]: new ShapeVisibilityActionProcessor(),
-    [ACTION_TYPE.STYLE]: new CommonStyleActionProcessor()
+    [ACTION_TYPE.STYLE]: new CommonStyleActionProcessor(),
+    [ACTION_TYPE.MOVETO]: new CommonMoveToActionProcessor()
   },
   [StoryComponentType.IMAGE]: {
     [ACTION_TYPE.APPEAR]: new ImageVisibilityActionProcessor(),
     [ACTION_TYPE.DISAPPEAR]: new ImageVisibilityActionProcessor(),
-    [ACTION_TYPE.STYLE]: new CommonStyleActionProcessor()
+    [ACTION_TYPE.STYLE]: new CommonStyleActionProcessor(),
+    [ACTION_TYPE.MOVETO]: new CommonMoveToActionProcessor()
   },
   [StoryComponentType.TEXT]: {
     [ACTION_TYPE.APPEAR]: new TextVisibilityActionProcessor(),
     [ACTION_TYPE.DISAPPEAR]: new TextVisibilityActionProcessor(),
-    [ACTION_TYPE.STYLE]: new CommonStyleActionProcessor()
+    [ACTION_TYPE.STYLE]: new CommonStyleActionProcessor(),
+    [ACTION_TYPE.MOVETO]: new CommonMoveToActionProcessor()
   },
   [StoryComponentType.TIMELINE]: {
     [ACTION_TYPE.APPEAR]: new TimelineVisibilityActionProcessor(),
     [ACTION_TYPE.DISAPPEAR]: new TimelineVisibilityActionProcessor(),
-    [ACTION_TYPE.STYLE]: new CommonStyleActionProcessor()
+    [ACTION_TYPE.STYLE]: new CommonStyleActionProcessor(),
+    [ACTION_TYPE.MOVETO]: new CommonMoveToActionProcessor()
   }
 };
