+++ packages/vstory/src/story/interface/dsl-interface.ts
@@ -2,10 +2,14 @@ import type {
   IChartVisibilityAction,
   IComponentVisibilityAction
 } from '../../player/processor/interface/appear-action';
-import type { IComponentStyleAction } from '../../player/processor/interface/style-action';
+import type { IComponentMoveToAction, IComponentStyleAction } from '../../player/processor/interface/style-action';
 import type { ICharacterSpec } from '../character';
 
-export type IAction = IComponentStyleAction | IComponentVisibilityAction | IChartVisibilityAction;
+export type IAction =
+  | IComponentStyleAction
+  | IComponentMoveToAction
+  | IComponentVisibilityAction
+  | IChartVisibilityAction;
 
 export interface IStorySpec {
   acts: IActSpec[]; // 作品的章节
